Implement physics lib

+++ libs/include/geometry/types.h
@@ -2,4 +2,4 @@
 
 #include "geometry/types-2d.h"
 
-using Mass = unsigned int;
+using Mass = float;
